Initial architecture rework of content card
